Permissions

+++ lib/GO/Modules/Contacts/Controller/ContactController.php
@@ -119,6 +119,9 @@ class ContactController extends AbstractController {
 	
 	protected function actionNew($returnAttributes = []){
 		$contact = new Contact();
+		if (!$contact->permissions->create) {
+			throw new Forbidden();
+		}
 		return $this->renderModel($contact, $returnAttributes);
 	}
 	
@@ -151,9 +154,9 @@ class ContactController extends AbstractController {
 		}
 
 
-//		if (!$contact->permission()->readAccess) {
-//			throw new Forbidden();
-//		}
+		if (!$contact->permissions->readAccess) {
+			throw new Forbidden();
+		}
 
 		return $this->renderModel($contact, $returnAttributes);
 
@@ -175,12 +178,12 @@ class ContactController extends AbstractController {
 	 */
 	public function actionCreate($returnAttributes = []) {
 		
+		$contact = new Contact();		
 		
-//		if (!ContactsModule::model()->checkPermission('createAccess')) {
-//			throw new Forbidden();
-//		}
-
-		$contact = new Contact();
+		if (!$contact->permissions->create) {
+			throw new Forbidden();
+		}
+		
 		$contact->setAttributes(App::request()->payload['data']);
 		$contact->save();
 		
@@ -216,9 +219,9 @@ class ContactController extends AbstractController {
 			throw new NotFound();
 		}
 		
-//		if (!$contact->checkPermission('editAccess')) {
-//			throw new Forbidden();
-//		}
+		if (!$contact->permissions->editAccess) {
+			throw new Forbidden();
+		}
 
 		$contact->setAttributes(App::request()->payload['data']);
 		$contact->save();
@@ -239,9 +242,9 @@ class ContactController extends AbstractController {
 			throw new NotFound();
 		}
 		
-//		if (!$contact->checkPermission('deleteAccess')) {
-//			throw new Forbidden();
-//		}
+		if (!$contact->permissions->deleteAccess) {
+			throw new Forbidden();
+		}
 
 		$contact->delete();
